requirements.txt
fastapi==0.110.1
pydantic-settings==2.2.1
jinja2==3.1.3
ujson==5.9.0
gunicorn==21.2.0
uvicorn==0.27.0.post1
redis[hiredis]==5.0.1
couchbase==4.2.0
psycopg2-binary==2.9.9
pytest==8.1.1
httpx==0.27.0
pyparsing==3.1.2
pandas==2.2.2
pillow==10.3.0
qrcode==7.4.2
